+++ ecs/src/v2/model/NovaCreateServersOption.cpp
@@ -13,6 +13,8 @@ namespace Model {
 
 NovaCreateServersOption::NovaCreateServersOption()
 {
+    autoTerminateTime_ = "";
+    autoTerminateTimeIsSet_ = false;
     imageRef_ = "";
     imageRefIsSet_ = false;
     flavorRef_ = "";
@@ -55,6 +57,9 @@ web::json::value NovaCreateServersOption::toJson() const
 {
     web::json::value val = web::json::value::object();
 
+    if(autoTerminateTimeIsSet_) {
+        val[utility::conversions::to_string_t("auto_terminate_time")] = ModelBase::toJson(autoTerminateTime_);
+    }
     if(imageRefIsSet_) {
         val[utility::conversions::to_string_t("imageRef")] = ModelBase::toJson(imageRef_);
     }
@@ -114,6 +119,15 @@ bool NovaCreateServersOption::fromJson(const web::json::value& val)
 {
     bool ok = true;
     
+    if(val.has_field(utility::conversions::to_string_t("auto_terminate_time"))) {
+        const web::json::value& fieldValue = val.at(utility::conversions::to_string_t("auto_terminate_time"));
+        if(!fieldValue.is_null())
+        {
+            std::string refVal;
+            ok &= ModelBase::fromJson(fieldValue, refVal);
+            setAutoTerminateTime(refVal);
+        }
+    }
     if(val.has_field(utility::conversions::to_string_t("imageRef"))) {
         const web::json::value& fieldValue = val.at(utility::conversions::to_string_t("imageRef"));
         if(!fieldValue.is_null())
@@ -270,6 +284,27 @@ bool NovaCreateServersOption::fromJson(const web::json::value& val)
     return ok;
 }
 
+
+std::string NovaCreateServersOption::getAutoTerminateTime() const
+{
+    return autoTerminateTime_;
+}
+
+void NovaCreateServersOption::setAutoTerminateTime(const std::string& value)
+{
+    autoTerminateTime_ = value;
+    autoTerminateTimeIsSet_ = true;
+}
+
+bool NovaCreateServersOption::autoTerminateTimeIsSet() const
+{
+    return autoTerminateTimeIsSet_;
+}
+
+void NovaCreateServersOption::unsetautoTerminateTime()
+{
+    autoTerminateTimeIsSet_ = false;
+}
 
 std::string NovaCreateServersOption::getImageRef() const
 {