+++ ecs/src/v2/model/UpdateAutoTerminateTimeServerResponse.cpp
@@ -1,0 +1,46 @@
+
+
+#include "huaweicloud/ecs/v2/model/UpdateAutoTerminateTimeServerResponse.h"
+
+namespace HuaweiCloud {
+namespace Sdk {
+namespace Ecs {
+namespace V2 {
+namespace Model {
+
+
+
+
+UpdateAutoTerminateTimeServerResponse::UpdateAutoTerminateTimeServerResponse()
+{
+}
+
+UpdateAutoTerminateTimeServerResponse::~UpdateAutoTerminateTimeServerResponse() = default;
+
+void UpdateAutoTerminateTimeServerResponse::validate()
+{
+}
+
+web::json::value UpdateAutoTerminateTimeServerResponse::toJson() const
+{
+    web::json::value val = web::json::value::object();
+
+
+    return val;
+}
+
+bool UpdateAutoTerminateTimeServerResponse::fromJson(const web::json::value& val)
+{
+    bool ok = true;
+    
+    return ok;
+}
+
+
+}
+}
+}
+}
+}
+
+