+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+# 3.0.3-beta 2021-01-25
+## HuaweiCloud SDK Core
+- ### Features
+    - None
+- ### Bug Fix
+    - Support using default config.
+- ### Change
+    - The default value of `ConnectionTimeout` is set to 60 seconds.
+    - The default value of `ReadTimeout` is set to 120 seconds.
+
+## HuaweiCloud SDK ECS
+- ### Features
+    - Support more interface: UpdateAutoTerminateTimeServer.
+- ### Bug Fix
+    - None
+- ### Change
+    - None
+
+
 # 3.0.2-beta 2021-01-15
 ## HuaweiCloud SDK Core
 - ### Features
@@ -9,7 +28,7 @@
 
 # 3.0.1-beta 2020-12-31
 ## First Release
- - ### Supported Services
+- ### Supported Services
     - `Elastic Cloud Server`(ECS) service.
     - `Elastic IP`(EIP) service.
     - `Virtual Private Cloud`(VPC) service.